simplified navbar and enabled chaning views
